pde-veracode-develop/python/requirements.txt
beautifulsoup4==4.10.0
boto3==1.20.45
botocore==1.23.45
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
docopt==0.6.2
httpie==1.0.3
idna==3.3
jmespath==0.10.0
lxml==4.6.4
markdownify==0.9.4
Pygments==2.10.0
python-dateutil==2.8.2
python-gitlab==2.10.1
requests==2.26.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3
strconv==0.4.2
tabulate==0.8.9
urllib3==1.26.7
veracode-api-signing==21.3.0
veracode-python==0.1.67
xmltodict==0.12.0
